+++ tests-mocha/IDBObjectStore/delete-spec.js
@@ -855,8 +855,8 @@ describe('IDBObjectStore.delete', function () {
                 err = e;
             }
 
-            expect(err).to.be.an.instanceOf(TypeError);
-            expect(err.name).to.equal('TypeError');
+            expect(err).to.be.an.instanceOf(env.DOMException);
+            expect(err.name).to.equal('DataError');
 
             db.close();
             done();